Move LX relayout planning under scratchpad
